Extract lens query cache helper
